server2/requirements.txt
gunicorn
flask
firebase_admin
databases
SpeechRecognition
pydub
moviepy
numpy
tensorflow==2.0
matplotlib
opencv-python
pyrebase4
